# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login user",
                "description": "Return user token",
                "operationId": "d618a255475240ba3d40cc39692cd924",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginUserSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponseSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example endpoint",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserSchema": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "user@gmail.com"
                    },
                    "last_login": {
                        "type": "string"
                    },
                    "is_suspended": {
                        "type": "boolean"
                    },
                    "token_valid_after": {
                        "type": "string"
                    },
                    "two_factor_status": {
                        "type": "boolean"
                    },
                    "verified": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginUserSchema": {
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@gmail.com"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginResponseSchema": {
                "properties": {
                    "token": {
                        "description": "token",
                        "type": "string",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            }
        }
    }
}